clean up gallery
