+++ src/components/config-gallery.js
@@ -1,8 +1,36 @@
 export const configGallery = [
   {
+    text: "Victory Area with Stroke",
+    slug: "victory-area-stroke",
+    code: require("!!raw!../screens/gallery/examples/victory-area-stroke.example.js")
+  }, {
+    text: "VictoryPortal with Stacked Area",
+    slug: "victory-portal-stacked-area",
+    code: require("!!raw!../screens/gallery/examples/victory-portal-stacked-area.example.js")
+  }, {
+    text: "Stream Graph with Custom Gradient Fill",
+    slug: "stream-graph",
+    code: require("!!raw!../screens/gallery/examples/stream-graph.example.js")
+  }, {
+    text: "100% Column Chart",
+    slug: "column-chart",
+    code: require("!!raw!../screens/gallery/examples/column-chart.example.js")
+  }, {
+    text: "Stacked Bars with Central Axis",
+    slug: "stacked-bars-central-axis",
+    code: require("!!raw!../screens/gallery/examples/stacked-bars-central-axis.example.js")
+  }, {
     text: "Alternative Events",
     slug: "alternative-events",
     code: require("!!raw!../screens/gallery/examples/alternative-events.example.js")
+  }, {
+    text: "Horizontal Grouped Bars",
+    slug: "horizontal-grouped-bars",
+    code: require("!!raw!../screens/gallery/examples/horizontal-grouped-bars.example.js")
+  }, {
+    text: "Stacked Grouped Bars",
+    slug: "stacked-grouped-bars",
+    code: require("!!raw!../screens/gallery/examples/stacked-grouped-bars.example.js")
   }, {
     text: "Animating Circular Progress Bar",
     slug: "animating-circular-progress-bar",
@@ -12,17 +40,9 @@ export const configGallery = [
     slug: "brush-zoom",
     code: require("!!raw!../screens/gallery/examples/brush-zoom.example.js")
   }, {
-    text: "100% Column Chart",
-    slug: "column-chart",
-    code: require("!!raw!../screens/gallery/examples/column-chart.example.js")
-  }, {
     text: "Custom Tooltip Labels",
     slug: "custom-tooltip-labels",
     code: require("!!raw!../screens/gallery/examples/custom-tooltip-labels.example.js")
-  }, {
-    text: "Horizontal Grouped Bars",
-    slug: "horizontal-grouped-bars",
-    code: require("!!raw!../screens/gallery/examples/horizontal-grouped-bars.example.js")
   }, {
     text: "Interpolation",
     slug: "interpolation",
@@ -32,29 +52,9 @@ export const configGallery = [
     slug: "multipoint-tooltip-labels",
     code: require("!!raw!../screens/gallery/examples/multipoint-tooltip-labels.example.js")
   }, {
-    text: "Simple Grouped Bars",
-    slug: "simple-grouped-bars",
-    code: require("!!raw!../screens/gallery/examples/simple-grouped-bars.example.js")
-  }, {
-    text: "Stacked Area Dates",
-    slug: "stacked-area-dates",
-    code: require("!!raw!../screens/gallery/examples/stacked-area-dates.example.js")
-  }, {
-    text: "Stacked Bars with Central Axis",
-    slug: "stacked-bars-central-axis",
-    code: require("!!raw!../screens/gallery/examples/stacked-bars-central-axis.example.js")
-  }, {
     text: "Stacked Polar Bars",
     slug: "stacked-polar-bars",
     code: require("!!raw!../screens/gallery/examples/stacked-polar-bars.example.js")
-  }, {
-    text: "Stream Graph with Custom Gradient Fill",
-    slug: "stream-graph",
-    code: require("!!raw!../screens/gallery/examples/stream-graph.example.js")
-  }, {
-    text: "Victory Area with Stroke",
-    slug: "victory-area-stroke",
-    code: require("!!raw!../screens/gallery/examples/victory-area-stroke.example.js")
   }, {
     text: "Victory Line with null data",
     slug: "victory-line-null-data",
@@ -63,10 +63,6 @@ export const configGallery = [
     text: "Victory Pie with Center Label",
     slug: "victory-pie-center-label",
     code: require("!!raw!../screens/gallery/examples/victory-pie-center-label.example.js")
-  }, {
-    text: "Victory Portal with Stacked Area",
-    slug: "victory-portal-stacked-area",
-    code: require("!!raw!../screens/gallery/examples/victory-portal-stacked-area.example.js")
   }, {
     text: "VictoryVoronoiContainer with Custom onClick",
     slug: "victory-voronoi-container-custom-on-click",
