+++ src/screens/gallery/examples/horizontal-grouped-bars.example.js
@@ -3,7 +3,7 @@
   will be removed before displaying this document to the user
 */
 /* global React, ReactDOM, App, mountNode */
-/* global VictoryChart, VictoryGroup, VictoryBar, VictoryTheme */
+/* global VictoryChart, VictoryGroup, VictoryBar */
 
 // Victory requires `react@^15.5.0` and `prop-types@^15.5.0`
 
@@ -12,15 +12,12 @@ class App extends React.Component {
     return (
       <div>
         <VictoryChart
-          theme={VictoryTheme.material}
           domain={{y: [0.5, 5.5]}}
         >
             <VictoryGroup horizontal
               offset={10}
               style={{ data: { width: 6 } }}
-              colorScale={
-                ["brown", "tomato", "gold"]
-              }
+              colorScale={["brown", "tomato", "gold"]}
             >
               <VictoryBar
                 data={[
